src/PythonScripts/printer/requirements.txt
flask
requests
websocket-client
asyncua
cryptography